+++ Box.V2/Managers/BoxTermsOfServiceManager.cs
@@ -3,6 +3,7 @@ using Box.V2.Config;
 using Box.V2.Converter;
 using Box.V2.Extensions;
 using Box.V2.Models;
+using Box.V2.Models.Request;
 using Box.V2.Services;
 using System;
 using System.Threading.Tasks;
@@ -106,11 +107,31 @@ namespace Box.V2.Managers
         /// </summary>
         /// <param name="termsOfServiceUserStatusesRequest">The request object for terms of service user status.</param>
         /// <returns>The status of the terms of service for a user.</returns>
+        [Obsolete("Use CreateBoxTermsOfServiceUserStatusesAsync(BoxTermsOfServiceStatusCreateRequest termsOfServiceUserStatusCreateRequest) instead.")]
         public async Task<BoxTermsOfServiceUserStatuses> CreateBoxTermsOfServiceUserStatusesAsync(BoxTermsOfServiceUserStatusesRequest termsOfServicesUserStatusesRequest)
         {
             BoxRequest request = new BoxRequest(_config.TermsOfServiceUserStatusesUri)
                 .Method(RequestMethod.Post)
                 .Payload(_converter.Serialize<BoxTermsOfServiceUserStatusesRequest>(termsOfServicesUserStatusesRequest));
+
+            IBoxResponse<BoxTermsOfServiceUserStatuses> response = await ToResponseAsync<BoxTermsOfServiceUserStatuses>(request).ConfigureAwait(false);
+
+            return response.ResponseObject;
+        }
+
+        /// <summary>
+        /// Create a terms of service status for user.
+        /// </summary>
+        /// <param name="termsOfServiceUserStatusCreateRequest">The request object for terms of service user status.</param>
+        /// <returns>The status of the terms of service for a user.</returns>
+        public async Task<BoxTermsOfServiceUserStatuses> CreateBoxTermsOfServiceUserStatusesAsync(BoxTermsOfServiceUserStatusCreateRequest termsOfServiceUserStatusCreateRequest)
+        {
+            termsOfServiceUserStatusCreateRequest.ThrowIfNull("TermsOfService");
+            termsOfServiceUserStatusCreateRequest.ThrowIfNull("User");
+
+            BoxRequest request = new BoxRequest(_config.TermsOfServiceUserStatusesUri)
+                .Method(RequestMethod.Post)
+                .Payload(_converter.Serialize<BoxTermsOfServiceUserStatusCreateRequest>(termsOfServiceUserStatusCreateRequest));
 
             IBoxResponse<BoxTermsOfServiceUserStatuses> response = await ToResponseAsync<BoxTermsOfServiceUserStatuses>(request).ConfigureAwait(false);